src\ja\yanmaga to DSL
