+++ src/ja/yanmaga/build.gradle.kts
@@ -4,13 +4,23 @@ plugins {
 
 keiyoushi {
     name = "Weekly Young Magazine"
-    className = "YanmagaFactory"
-    versionCode = 4
+    versionCode = 5
     contentWarning = ContentWarning.NSFW
     libVersion = "1.4"
+
+    source {
+        name = "ヤンマガ（マンガ）"
+        lang = "ja"
+        baseUrl = "https://yanmaga.jp"
+    }
+
+    source {
+        name = "ヤンマガ（グラビア）"
+        lang = "ja"
+        baseUrl = "https://yanmaga.jp"
+    }
 }
 
 dependencies {
-
     implementation(project(":lib:speedbinb"))
 }
