--- src/ja/yanmaga/src/eu/kanade/tachiyomi/extension/ja/yanmaga/YanmagaFactory.kt
@@ -1,10 +1,0 @@
-package eu.kanade.tachiyomi.extension.ja.yanmaga
-
-import eu.kanade.tachiyomi.source.SourceFactory
-
-class YanmagaFactory : SourceFactory {
-    override fun createSources() = listOf(
-        YanmagaComics(),
-        YanmagaGravures(),
-    )
-}
